move MovementHandler into separate file
